updated member portal module

+++ resources/views/quizzes/list.blade.php
@@ -2,7 +2,7 @@
     <x-slot name="header">
         <div class="flex justify-between"> 
             <h2 class="font-semibold text-4xl text-white dark:text-gray-200 leading-tight">
-                Quizzes     
+                {{ __('Quizzes') }}
             </h2>
             @can('create quizzes')
             <a href="{{ route('quizzes.create') }}" class="inline-block px-5 py-2 text-white hover:text-[#101966] hover:border-[#101966] bg-[#101966] hover:bg-white focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-[#101966] border border-white border font-medium dark:border-[#3E3E3A] dark:hover:bg-black dark:hover:border-[#3F53E8] rounded-lg text-xl leading-normal">Create</a>
@@ -19,11 +19,12 @@
                     <table id="quizzesTable" class="w-full">
                         <thead class="bg-gray-50">
                             <tr class="border-b">
-                                <th class="px-6 py-3 text-center">#</th>
-                                <th class="px-6 py-3 text-center">Title</th>
-                                <th class="px-6 py-3 text-center">Description</th>
-                                <th class="px-6 py-3 text-center">Author</th>
-                                <th class="px-6 py-3 text-center">Created</th>
+                                <th class="px-6 py-3 text-left">#</th>
+                                <th class="px-6 py-3 text-left">Title</th>
+                                <th class="px-6 py-3 text-left">Status</th>
+                                <th class="px-6 py-3 text-left">Responses</th>
+                                <th class="px-6 py-3 text-left">Author</th>
+                                <th class="px-6 py-3 text-left">Created</th>
                                 <th class="px-6 py-3 text-center">Action</th>
                             </tr>
                         </thead>
@@ -45,23 +46,23 @@
                     serverSide: true,
                     ajax: "{{ route('quizzes.index') }}",
                     columns: [
-                        { data: 'DT_RowIndex', name: 'DT_RowIndex', orderable: false, searchable: false, className: 'text-center'},
-                        { data: 'title', name: 'title', className: 'text-center'},
-                        { data: 'description', name: 'description', className: 'text-center' },
-                        { data: 'author', name: 'user.first_name', className: 'text-center' },
-                        { data: 'created_at', name: 'created_at', className: 'text-center' },
-                        { data: 'action', name: 'action', orderable: false, searchable: false, className: 'text-center' }
+                        { data: 'DT_RowIndex', name: 'DT_RowIndex', orderable: false, searchable: false },
+                        { data: 'title', name: 'title' },
+                        { data: 'is_published', name: 'is_published' },
+                        { data: 'responses_count', name: 'responses_count', searchable: false },
+                        { data: 'author', name: 'user.first_name' },
+                        { data: 'created_at', name: 'created_at' },
+                        { data: 'action', name: 'action', orderable: false, searchable: false }
                     ],
                     responsive: true,
                     autoWidth: false,
                     lengthMenu: [10, 25, 50, 100],
-                    pageLength: 10,
-                    order: [[4, 'desc']]
+                    pageLength: 10
                 });
             });
 
             function deleteQuiz(id) {
-                if (confirm("Are you sure you want to delete this quiz?")) {
+                if (confirm("Are you sure you want to delete this quiz and all its responses?")) {
                     $.ajax({
                         url: '{{ route("quizzes.destroy") }}',
                         type: 'delete',
